SUCCESS!!!
I will be removing matplotlib to handle the 3D stuff. It is too slow to handle the number of triangles for the many parts that are going to be used.
As an example, a transition is going to have roughly 20K triangles to it and that's just a single transition. get into 3D models for connectors and the wires/bundles and that number is going to get very large very fast. Need to use the hardware acceleration on a video card to handle the millions of triangles that is going to make up a harness.